refactor query dto update service
