+++ apps/api/src/donation-wish/donation-wish.controller.ts
@@ -3,7 +3,7 @@ import { Public } from 'nest-keycloak-connect'
 import { DonationWishService } from './donation-wish.service'
 import { CreateDonationWishDto } from './dto/create-donation-wish.dto'
 import { ApiTags, ApiQuery } from '@nestjs/swagger'
-import { PagingQueryDto } from '../common/dto/paging-query-dto'
+import { PagingQueryDto } from '../common/dto/donation-query-dto'
 
 @ApiTags('donation-wish')
 @Controller('donation-wish')
